src/predict/score.py
# TODO: Create AzureML scoring script for real-time endpoint
# - Define init() function to load model from AzureML registry
# - Define run() function to receive JSON input and return predictions
# - Add input validation and error handling
# - Return prediction probabilities and risk flags
# - Add logging for monitoring
# - Ensure compatibility with AzureML Managed Endpoint

import json
import joblib
from azureml.core.model import Model

# Add your scoring logic here
