requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
imbalanced-learn==0.11.0

# MLflow (removed Azure dependencies - using local MLflow)
mlflow==2.16.2
sqlalchemy>=2.0.0

# Data Warehouse Connectivity
dbt-core>=1.7.0
dbt-postgres>=1.7.0
psycopg2-binary>=2.9.9

# Airflow & Orchestration
apache-airflow>=2.7.3
apache-airflow-providers-postgres>=5.7.0
apache-airflow-providers-microsoft-azure>=8.4.0

# API & Data Validation
fastapi==0.109.0
pydantic==2.5.3
uvicorn==0.25.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
pyarrow==14.0.2
